Align Worker with Laravel
- Add WorkerStarting event dispatch
- Add stop reasons (WorkerStopReason) to stopIfNecessary/stop
- Add pause/resume polling via queuePaused()
- Add $restartable, $pausable, $memoryExceededExitCode static properties
- Add SIGINT signal handler
- Add memoryLimit > 0 guard
- Pass queue name to JobPopping event
- Pass exception object (not bool) to JobAttempted event
- Pass backoff to JobReleasedAfterException event
- Add stop check in concurrency-full branch
- Change cache type from CacheFactory to CacheContract (Repository)
- Use hypervel:queue:restart cache key constant

+++ src/queue/src/WorkerOptions.php
@@ -11,10 +11,10 @@ class WorkerOptions
      *
      * @param string $name the name of the worker
      * @param int|int[] $backoff
-     * @param int $memory the number of seconds to wait before retrying a job that encountered an uncaught exception
-     * @param int $timeout the maximum amount of RAM the worker may consume
+     * @param float $memory the maximum amount of RAM the worker may consume
+     * @param int $timeout the maximum number of seconds a child worker may run
      * @param int $sleep the number of seconds to wait in between polling the queue
-     * @param int $maxTries the number of seconds to rest between jobs
+     * @param int $maxTries the maximum number of times a job may be attempted
      * @param bool $force indicates if the worker should run in maintenance mode
      * @param bool $stopWhenEmpty indicates if the worker should stop when the queue is empty
      * @param int $maxJobs the maximum number of jobs to run
@@ -25,7 +25,7 @@ class WorkerOptions
     public function __construct(
         public string $name = 'default',
         public array|int $backoff = 0,
-        public int $memory = 128,
+        public float $memory = 128,
         public int $timeout = 60,
         public int $sleep = 3,
         public int $maxTries = 1,
